* applied many compiler warning recommendations * cleaned up code * added unit test code * migrated ranking RCI computation to kelondroFlex and kelondroCollectionIndex
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@2414 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/plasma/plasmaRankingCRProcess.java
@@ -52,6 +52,11 @@ import java.util.Map;
 
 import de.anomic.kelondro.kelondroAttrSeq;
 import de.anomic.kelondro.kelondroBase64Order;
+import de.anomic.kelondro.kelondroCollectionIndex;
+import de.anomic.kelondro.kelondroFlexTable;
+import de.anomic.kelondro.kelondroIndex;
+import de.anomic.kelondro.kelondroRow;
+import de.anomic.kelondro.kelondroRowSet;
 import de.anomic.server.serverFileUtils;
 import de.anomic.server.serverDate;
 import de.anomic.tools.bitfield;
@@ -63,7 +68,20 @@ public class plasmaRankingCRProcess {
     header.append("# Created=" + System.currentTimeMillis()); header.append((char) 13); header.append((char) 10);
     header.append("# Structure=<Referee-12>,'=',<UDate-3>,<VDate-3>,<LCount-2>,<GCount-2>,<ICount-2>,<DCount-2>,<TLength-3>,<WACount-3>,<WUCount-3>,<Flags-1>,'|',*<Anchor-" + ((type.equals("crl")) ? "6" : "12") + ">"); header.append((char) 13); header.append((char) 10);
     header.append("# ---"); header.append((char) 13); header.append((char) 10);
-    */  
+    */
+    
+    public static final kelondroRow CRG_accrow = new kelondroRow(
+            "byte[] Referee-12," +
+            "Cardinal UDate-3 {b64e}, Cardinal VDate-3 {b64e}, " +
+            "Cardinal LCount-2 {b64e}, Cardinal GCount-2 {b64e}, Cardinal ICount-2 {b64e}, Cardinal DCount-2 {b64e}, Cardinal TLength-3 {b64e}, " +
+            "Cardinal WACount-3 {b64e}, Cardinal WUCount-3 {b64e}, Cardinal Flags-1 {b64e}, " +
+            "Cardinal FUDate-3 {b64e}, Cardinal FDDate-3 {b64e}, Cardinal LUDate-3 {b64e}, " + 
+            "Cardinal UCount-2 {b64e}, Cardinal PCount-2 {b64e}, Cardinal ACount-2 {b64e}, Cardinal VCount-2 {b64e}, Cardinal Vita-2 {b64e}");
+    public static final kelondroRow CRG_colrow = new kelondroRow("byte[] Anchor-12");
+    public static final String CRG_accname = "CRG-a-attr";
+    public static final String CRG_seqname = "CRG-a-coli";
+    public static final kelondroRow RCI_coli = new kelondroRow("byte[] RefereeDom-6");
+    public static final String RCI_colname = "RCI-a-coli";
 
     private static boolean accumulate_upd(File f, kelondroAttrSeq acc) {
         // open file
@@ -96,7 +114,7 @@ public class plasmaRankingCRProcess {
                 Vita   = (int) acc_entry.getAttr("Vita", 0);
                 
                 // update counters and dates
-                acc_entry.setSeq(new_entry.getSeq()); // need to be checked
+                acc_entry.setSeq(new_entry.getSeqSet()); // need to be checked
                 
                 UCount++; // increase update counter
                 PCount += (new_flags.get(1)) ? 1 : 0;
@@ -111,7 +129,7 @@ public class plasmaRankingCRProcess {
                 acc_entry.setAttr("Flags", (int) kelondroBase64Order.enhancedCoder.decodeLong(new String(acc_flags.getBytes())));
             } else {
                 // initialize counters and dates
-                acc_entry = acc.newEntry(key, new_entry.getAttrs(), new_entry.getSeq());
+                acc_entry = acc.newEntry(key, new_entry.getAttrs(), new_entry.getSeqSet());
                 FUDate = plasmaWordIndex.microDateHoursInt(System.currentTimeMillis()); // first update date
                 FDDate = plasmaWordIndex.microDateHoursInt(System.currentTimeMillis()); // very difficult to compute; this is only a quick-hack
                 LUDate = (int) new_entry.getAttr("VDate", 0);
@@ -138,7 +156,86 @@ public class plasmaRankingCRProcess {
         return true;
     }
     
-    public static void accumulate(File from_dir, File tmp_dir, File err_dir, File bkp_dir, File to_file, int max_files) throws IOException {
+    private static boolean accumulate_upd(File f, kelondroIndex acc, kelondroCollectionIndex seq) throws IOException {
+        // open file
+        kelondroAttrSeq source_cr = null;
+        try {
+            source_cr = new kelondroAttrSeq(f, false);
+        } catch (IOException e) {
+            return false;
+        }
+        
+        // put elements in accumulator file
+        Iterator el = source_cr.keys();
+        String key;
+        kelondroAttrSeq.Entry new_entry;
+        kelondroRow.Entry acc_entry;
+        int FUDate, FDDate, LUDate, UCount, PCount, ACount, VCount, Vita;
+        bitfield acc_flags, new_flags;
+        while (el.hasNext()) {
+            key = (String) el.next();
+            new_entry = source_cr.getEntry(key);
+            new_flags = new bitfield(kelondroBase64Order.enhancedCoder.encodeLong((long) new_entry.getAttr("Flags", 0), 1).getBytes());
+            // enrich information with additional values
+            if ((acc_entry = acc.get(key.getBytes())) != null) {
+                FUDate = (int) acc_entry.getColLong("FUDate", 0);
+                FDDate = (int) acc_entry.getColLong("FDDate", 0);
+                LUDate = (int) acc_entry.getColLong("LUDate", 0);
+                UCount = (int) acc_entry.getColLong("UCount", 0);
+                PCount = (int) acc_entry.getColLong("PCount", 0);
+                ACount = (int) acc_entry.getColLong("ACount", 0);
+                VCount = (int) acc_entry.getColLong("VCount", 0);
+                Vita   = (int) acc_entry.getColLong("Vita", 0);
+                
+                // update counters and dates
+                seq.put(key.getBytes(), new_entry.getSeqCollection()); // FIXME: old and new collection must be joined
+                
+                UCount++; // increase update counter
+                PCount += (new_flags.get(1)) ? 1 : 0;
+                ACount += (new_flags.get(2)) ? 1 : 0;
+                VCount += (new_flags.get(3)) ? 1 : 0;
+                
+                // 'OR' the flags
+                acc_flags = new bitfield(kelondroBase64Order.enhancedCoder.encodeLong(acc_entry.getColLong("Flags", 0), 1).getBytes());
+                for (int i = 0; i < 6; i++) {
+                    if (new_flags.get(i)) acc_flags.set(i, true);
+                }
+                acc_entry.setCol("Flags", (int) kelondroBase64Order.enhancedCoder.decodeLong(new String(acc_flags.getBytes())));
+            } else {
+                // initialize counters and dates
+                acc_entry = acc.row().newEntry();
+                acc_entry.setCol("Referee", key, null);
+                for (int i = 1; i < acc.row().columns(); i++) {
+                    acc_entry.setCol(i, new_entry.getAttr(acc.row().column(i).nickname(), 0));
+                }
+                seq.put(key.getBytes(), new_entry.getSeqCollection());
+                FUDate = plasmaWordIndex.microDateHoursInt(System.currentTimeMillis()); // first update date
+                FDDate = plasmaWordIndex.microDateHoursInt(System.currentTimeMillis()); // very difficult to compute; this is only a quick-hack
+                LUDate = (int) new_entry.getAttr("VDate", 0);
+                UCount = 0;
+                PCount = (new_flags.get(1)) ? 1 : 0;
+                ACount = (new_flags.get(2)) ? 1 : 0;
+                VCount = (new_flags.get(3)) ? 1 : 0;
+                Vita   = 0;
+            }
+            // make plausibility check?
+            
+            // insert into accumulator
+            acc_entry.setCol("FUDate", (long) FUDate);
+            acc_entry.setCol("FDDate", (long) FDDate);
+            acc_entry.setCol("LUDate", (long) LUDate);
+            acc_entry.setCol("UCount", (long) UCount);
+            acc_entry.setCol("PCount", (long) PCount);
+            acc_entry.setCol("ACount", (long) ACount);
+            acc_entry.setCol("VCount", (long) VCount);
+            acc_entry.setCol("Vita", (long) Vita);
+            acc.put(acc_entry);
+        }
+        
+        return true;
+    }
+    
+    public static void accumulate(File from_dir, File tmp_dir, File err_dir, File bkp_dir, File to_file, int max_files, boolean newdb) throws IOException {
         if (!(from_dir.isDirectory())) {
             System.out.println("source path " + from_dir + " is not a directory.");
             return;
@@ -158,16 +255,23 @@ public class plasmaRankingCRProcess {
         
         // open target file
         kelondroAttrSeq acc = null;
-        if (!(to_file.exists())) {
-            acc = new kelondroAttrSeq("Global Ranking Accumulator File",
+        kelondroIndex newacc = null;
+        kelondroCollectionIndex newseq = null;
+        if (newdb) {
+            File path = to_file.getParentFile(); // path to storage place
+            newacc = new kelondroFlexTable(path, CRG_accname, kelondroBase64Order.enhancedCoder, 128 * 1024 * 1024, -1, CRG_accrow, true);
+            newseq = new kelondroCollectionIndex(path, CRG_seqname, 12, kelondroBase64Order.enhancedCoder, 128 * 1024 * 1024, -1, 2, CRG_colrow);
+        } else {
+            if (!(to_file.exists())) {
+                acc = new kelondroAttrSeq("Global Ranking Accumulator File",
                     "<Referee-12>,'='," +
                     "<UDate-3>,<VDate-3>,<LCount-2>,<GCount-2>,<ICount-2>,<DCount-2>,<TLength-3>,<WACount-3>,<WUCount-3>,<Flags-1>," +
                     "<FUDate-3>,<FDDate-3>,<LUDate-3>,<UCount-2>,<PCount-2>,<ACount-2>,<VCount-2>,<Vita-2>," +
                     "'|',*<Anchor-12>", false);
-            acc.toFile(to_file);
-        }
-        acc = new kelondroAttrSeq(to_file, false);
-        
+                acc.toFile(to_file);
+            }
+            acc = new kelondroAttrSeq(to_file, false);
+        }        
         // collect source files
         File source_file = null;
         String[] files = from_dir.list();
@@ -175,30 +279,47 @@ public class plasmaRankingCRProcess {
         for (int i = 0; i < max_files; i++) {
             // open file
             source_file = new File(from_dir, files[i]);
-            if (accumulate_upd(source_file, acc)) {
-                // move cr file to temporary folder
-                source_file.renameTo(new File(tmp_dir, files[i]));
+            if (newdb) {
+                if (accumulate_upd(source_file, newacc, newseq)) {
+                    // move cr file to temporary folder
+                    source_file.renameTo(new File(tmp_dir, files[i]));
+                } else {
+                    // error case: the cr-file is not valid; move to error path
+                    source_file.renameTo(new File(err_dir, files[i]));
+                }
             } else {
-                // error case: the cr-file is not valid; move to error path
-                source_file.renameTo(new File(err_dir, files[i]));
+                if (accumulate_upd(source_file, acc)) {
+                    // move cr file to temporary folder
+                    source_file.renameTo(new File(tmp_dir, files[i]));
+                } else {
+                    // error case: the cr-file is not valid; move to error path
+                    source_file.renameTo(new File(err_dir, files[i]));
+                }
             }
         }
         
-        // save accumulator to temporary file
-        File tmp_file;
-        if (to_file.toString().endsWith(".gz")) {
-            tmp_file = new File(to_file.toString() + "." + (System.currentTimeMillis() % 1000) + ".tmp.gz");
-        } else {
-            tmp_file = new File(to_file.toString() + "." + (System.currentTimeMillis() % 1000) + ".tmp");
-        }
         try {
-            acc.toFile(tmp_file);
-            // since this was successful, we remove the old file and move the new file to it
-            to_file.delete();
-            tmp_file.renameTo(to_file);
+            if (newdb) {
+                newacc.close();
+                newseq.close();
+            } else {
+                // save accumulator to temporary file
+                File tmp_file;
+                if (to_file.toString().endsWith(".gz")) {
+                    tmp_file = new File(to_file.toString() + "." + (System.currentTimeMillis() % 1000) + ".tmp.gz");
+                } else {
+                    tmp_file = new File(to_file.toString() + "." + (System.currentTimeMillis() % 1000) + ".tmp");
+                }
+                // store the file
+                acc.toFile(tmp_file);
+                // since this was successful, we remove the old file and move the new file to it
+                to_file.delete();
+                tmp_file.renameTo(to_file);
+            }
             serverFileUtils.moveAll(tmp_dir, bkp_dir);
         } catch (IOException e) {
             // move previously processed files back
+            e.printStackTrace();
             serverFileUtils.moveAll(tmp_dir, from_dir);
         }
         
@@ -232,7 +353,7 @@ public class plasmaRankingCRProcess {
             cr_UDate = cr_entry.getAttr("UDate", 0);
             
             // loop over all anchors
-            Iterator j = cr_entry.getSeq().entrySet().iterator();
+            Iterator j = cr_entry.getSeqSet().iterator();
             Map.Entry entry;
             while (j.hasNext()) {
                 // get domain of anchors
@@ -243,7 +364,7 @@ public class plasmaRankingCRProcess {
                 // update domain-specific entry
                 rci_entry = rci.getEntry(anchorDom);
                 if (rci_entry == null) rci_entry = rci.newEntry(anchorDom, false);
-                rci_entry.addSeq(referee, null);
+                rci_entry.addSeq(referee);
                 
                 // update Update-Date
                 rci_UDate = rci_entry.getAttr("UDate", 0);
@@ -268,11 +389,61 @@ public class plasmaRankingCRProcess {
         return count;
     }
     
+    public static int genrcix(File cr_path_in, File rci_path_out) throws IOException {
+        //kelondroFlexTable       acc = new kelondroFlexTable(cr_path_in, CRG_accname, kelondroBase64Order.enhancedCoder, 128 * 1024 * 1024, -1, CRG_accrow, true);
+        kelondroCollectionIndex seq = new kelondroCollectionIndex(cr_path_in, CRG_seqname, 12, kelondroBase64Order.enhancedCoder, 128 * 1024 * 1024, -1, 2, CRG_colrow);
+        kelondroCollectionIndex rci = new kelondroCollectionIndex(rci_path_out, RCI_colname, 6, kelondroBase64Order.enhancedCoder, 128 * 1024 * 1024, -1, 2, RCI_coli);
+        
+        // loop over all referees
+        int count = 0;
+        int size = seq.size();
+        long start = System.currentTimeMillis();
+        long l;
+        final Iterator i = seq.keycollections(null, false);
+        Object[] keycollection;
+        String referee, refereeDom, anchor, anchorDom;
+        kelondroRowSet cr_entry, rci_entry;
+        while (i.hasNext()) {
+            keycollection = (Object[]) i.next();
+            referee = new String((byte[]) keycollection[0]);
+            if (referee.length() == 6) refereeDom = referee; else refereeDom = referee.substring(6);
+            cr_entry = (kelondroRowSet) keycollection[1];
+            
+            // loop over all anchors
+            Iterator j = cr_entry.rows();
+            kelondroRow.Entry entry;
+            while (j.hasNext()) {
+                // get domain of anchors
+                entry = (kelondroRow.Entry) j.next();
+                anchor = (String) entry.getColString(0, null);
+                if (anchor.length() == 6) anchorDom = anchor; else anchorDom = anchor.substring(6);
+
+                // update domain-specific entry
+                rci_entry = rci.get(anchorDom.getBytes(), false);
+                if (rci_entry == null) rci_entry = new kelondroRowSet(RCI_coli);
+                rci_entry.add(refereeDom.getBytes());
+                
+                // insert entry
+                rci.put(anchorDom.getBytes(), rci_entry);
+            }
+            count++;
+            if ((count % 1000) == 0) {
+                l = java.lang.Math.max(1, (System.currentTimeMillis() - start) / 1000);
+                System.out.println("processed " + count + " citations, " + (count / l) + " per second, rci.size = " + rci.size() + ", " + ((size - count) / (count / l) / 60) + " minutes remaining; mem = " + Runtime.getRuntime().freeMemory());
+            }
+        }
+
+        // finished. write to file
+        seq.close();
+        rci.close();
+        return count;
+    }
+    
     public static void main(String[] args) {
         // java -classpath source de.anomic.plasma.kelondroPropFile -transcode DATA/RANKING/GLOBAL/CRG-test-unsorted-original.cr DATA/RANKING/GLOBAL/CRG-test-generated.cr
         try {
             if ((args.length == 5) && (args[0].equals("-accumulate"))) {
-                accumulate(new File(args[1]), new File(args[2]), new File(args[3]), new File(args[4]), new File(args[5]), Integer.parseInt(args[6]));
+                accumulate(new File(args[1]), new File(args[2]), new File(args[3]), new File(args[4]), new File(args[5]), Integer.parseInt(args[6]), true);
             }
             if ((args.length == 2) && (args[0].equals("-accumulate"))) {
                 File root_path = new File(args[1]);
@@ -291,7 +462,7 @@ public class plasmaRankingCRProcess {
                 serverFileUtils.moveAll(from_dir, ready_dir);
                 long start = System.currentTimeMillis();
                 int files = ready_dir.list().length;
-                accumulate(ready_dir, tmp_dir, err_dir, acc_dir, to_file, 1000);
+                accumulate(ready_dir, tmp_dir, err_dir, acc_dir, to_file, 1000, true);
                 long seconds = java.lang.Math.max(1, (System.currentTimeMillis() - start) / 1000);
                 System.out.println("Finished accumulate for " + files + " files in " + seconds + " seconds (" + (files / seconds) + " files/second)");
             }
@@ -335,6 +506,17 @@ public class plasmaRankingCRProcess {
             if ((args.length == 2) && (args[0].equals("-genrci"))) {
                 File root_path = new File(args[1]);
                 File cr_filedir = new File(root_path, "DATA/RANKING/GLOBAL/020_con0");
+                File rci_filedir = new File(root_path, "DATA/RANKING/GLOBAL/030_rci0");
+                rci_filedir.mkdirs();
+                long start = System.currentTimeMillis();
+                int count = genrcix(cr_filedir, rci_filedir);
+                long seconds = java.lang.Math.max(1, (System.currentTimeMillis() - start) / 1000);
+                System.out.println("Completed RCI generation: " + count + " citation references in " + seconds + " seconds (" + (count / seconds) + " CR-records/second)");                
+            }
+            /*
+            if ((args.length == 2) && (args[0].equals("-genrci"))) {
+                File root_path = new File(args[1]);
+                File cr_filedir = new File(root_path, "DATA/RANKING/GLOBAL/020_con0");
                 File rci_file = new File(root_path, "DATA/RANKING/GLOBAL/030_rci0/RCI-0.rci.gz");
                 rci_file.getParentFile().mkdirs();
                 String[] cr_filenames = cr_filedir.list();
@@ -345,6 +527,7 @@ public class plasmaRankingCRProcess {
                     System.out.println("Completed RCI generation for input file " + cr_filenames[i] + ": " + count + " citation references in " + seconds + " seconds (" + (count / seconds) + " CR-records/second)");
                 }
             }
+            */
         } catch (IOException e) {
             e.printStackTrace();
         }
